Extracted Local File Helper.
