+++ lib/MochaTest.js
@@ -1,7 +1,7 @@
-const readDir = require('./utils/rreadDir');
 const fs = require('fs');
 const Mocha = require('./../node_modules/mocha');
 const axios = require('./../node_modules/axios');
+const localHelper = require('./localFileHelper');
 
 const tempFilePath = './../src/temp/temp.js';
 const mochaTestsPath = './../src/mochaTests/';
@@ -20,8 +20,8 @@ class MochaTest {
     async runTest(){
         const studentFiles = await this.getStudentFile();
         if(!studentFiles) return;
-        writeTempFile(tempFilePath, studentFiles);
-        const mochaTestFile = await getTestFilePath(this.testFile);
+        localHelper.writeTempFile(tempFilePath, studentFiles);
+        const mochaTestFile = await await localHelper.getFilePath(mochaTestsPath, this.testFile);
         runMochaTest(mochaTestFile);
     }
 
@@ -112,27 +112,9 @@ function runMochaTest(testFilePath){
     mocha.run()
         .on('end', function() {
             console.log('All done');
-            deleteTempFile(tempFilePath)
+            localHelper.deleteTempFile(tempFilePath)
         });
 }
-
-async function getTestFilePath(fileName) {
-    return await getFilePath(mochaTestsPath, fileName);
-}
-
-function deleteTempFile(path) {
-    fs.unlink(path, err => {
-        if(err){
-            console.error(`Failed to delete temp file:
-${err}`);
-        }
-    })
-}
-
-function writeTempFile(path, string){
-    fs.writeFileSync(path, string);
-}
-
 
 async function getlocalStudentFile(name, filepath){
     let f = await getLocalStudentFilePaths(name, filepath);
@@ -158,16 +140,7 @@ Make sure you're using the full file path and try again.`)
 
 async function getLocalStudentFilePath(name){
     const filepath = process.cwd();
-    return await getFilePath(filepath, name);
-}
-
-async function getFilePath(baseDirectory, fileName) {
-    let directory = await readDir(baseDirectory);
-    let d = directory.filter(f => isJsFile(f) && getFileName(f) === fileName);
-    if(!d)
-        console.log(`A file name ${name}.js could not be found
-Make sure you're in the correct directory and try again`);
-    return d.shift();
+    return await localHelper.getFilePath(filepath, name);
 }
 
 const composeFile = (extra, dom, base, functions) => `${extra}
@@ -178,9 +151,7 @@ ${getFunctions(functions)}`;
 const getFunctions = functions => `
 module.exports = { ${functions.reduce((acc, func) => acc + `${func}: (typeof ${func}) === 'function' && ${func},`, '')} }`;
 
-const isJsFile = path => path.split('.').pop() === 'js';
 
-const getFileName = path => path.replace(/^.*[\\\/]/, '').split('.').shift();
 
 
 module.exports = MochaTest;
