+++ lib/TestCafeTest.js
@@ -28,8 +28,9 @@ class TestCafeTest {
 function runTestCafeTest(tempTestPath) {
     exec(`../node_modules/.bin/testcafe chrome ${tempTestPath} --colors --stop-on-first-fail`, (error, stdout, stderr) =>{
         console.log(stdout || stderr);
-        exec(`rm ${tempFilePath}`);
+        deleteTempFile(tempFilePath);
     });
+
 }
 
 function getTestCafeTest(path, fixtureName, URL) {
@@ -43,6 +44,14 @@ function getTestCafeTest(path, fixtureName, URL) {
 }
 
 
+function deleteTempFile(path) {
+    fs.unlink(path, err => {
+        if(err){
+            console.error(`Failed to delete temp file:
+${err}`);
+        }
+    })
+}
 async function getFilePath(baseDirectory, fileName) {
     let directory = await readDir(baseDirectory);
     let d = directory.filter(f => isJsFile(f) && getFileName(f) === fileName);
